fix mkldnn_rnn module import issue
